+++ changelog.md
@@ -3,3 +3,4 @@
 more data than they should (#434)
 - Fixed oxygen distributor displaying an incorrect oxygen consumption rate on fabric
 - Re-added the disabled planets config option
+- Fixed mixin crash with some mod (#431)